rpi5b/mqtt-led/requirements.txt
paho-mqtt>=1.6
gpiozero>=2.0
lgpio>=0.2
